+++ go/enclave/rpc/rpc_utils.go
@@ -6,8 +6,6 @@ import (
 
 	"github.com/ethereum/go-ethereum/common/hexutil"
 
-	"github.com/obscuronet/go-obscuro/go/common"
-
 	gethcommon "github.com/ethereum/go-ethereum/common"
 )
 
@@ -28,21 +26,6 @@ func ExtractTxHash(getTxReceiptParams []byte) (gethcommon.Hash, error) {
 	}
 	txHash := gethcommon.HexToHash(paramsJSONList[0]) // The only argument is the transaction hash.
 	return txHash, err
-}
-
-// ExtractTx returns the common.L2Tx from the params of an eth_sendRawTransaction request.
-func ExtractTx(sendRawTxParams []byte) (*common.L2Tx, error) {
-	// We need to extract the transaction hex from the JSON list encoding. We remove the leading `"[0x`, and the trailing `]"`.
-	txBinary := sendRawTxParams[4 : len(sendRawTxParams)-2]
-	txBytes := gethcommon.Hex2Bytes(string(txBinary))
-
-	tx := &common.L2Tx{}
-	err := tx.UnmarshalBinary(txBytes)
-	if err != nil {
-		return nil, fmt.Errorf("could not unmarshall transaction from bytes. Cause: %w", err)
-	}
-
-	return tx, nil
 }
 
 // ExtractCallParamTo extracts and parses the `to` field of an eth_call request.
